+++ app/post/post-module.js
@@ -11,11 +11,11 @@ angular.module('ushahidi.posts', [])
 .directive('postMediaValue', require('./directives/post-media-value-directive.js'))
 .directive('postPreviewMedia', require('./directives/post-preview-media-directive.js'))
 // Post editing workflows
-.directive('postEditor', require('./directives/post-editor-directive.js'))
+.directive('postEditor', require('./directives/post-editor.directive.js'))
 .directive('postChooseForm', require('./directives/post-choose-form-directive.js'))
-.directive('postValueEdit', require('./directives/post-value-edit-directive.js'))
-.directive('postStatus', require('./directives/post-status-directive.js'))
-.directive('postStages', require('./directives/post-stages-directive.js'))
+.directive('postValueEdit', require('./directives/post-value-edit.directive.js'))
+.directive('postStatus', require('./directives/post-status.directive.js'))
+.directive('postVerticalTabs', require('./directives/post-vertical-tabs.directive.js'))
 // Views
 .directive('postView', require('./directives/views/post-view-directive.js'))
 .directive('postViewMap', require('./directives/views/post-view-map-directive.js'))
