+++ app/post/post-routes.js
@@ -33,7 +33,7 @@ function (
     })
     .when('/posts/:id/edit', {
         controller: require('./controllers/post-edit-controller.js'),
-        templateUrl: 'templates/posts/modify.html',
+        templateUrl: 'templates/posts/modify/modify.html',
         resolve: {
             post: ['$route', 'PostEndpoint', function ($route, PostEndpoint) {
                 return PostEndpoint.get({id: $route.current.params.id}).$promise;
