+++ server/www/templates/common/publish-selector/publish-selector.html
@@ -2,11 +2,11 @@
     <legend data-toggle="dropdown-menu" class="dropdown-trigger init" dropdown-toggle translate>post.who_can_see_this</legend>
     <div class="dropdown-menu init" dropdown-menu>
         <div class="form-fieldgroup">
-            <div class="form-field radio">
-                <input 
-                    name="{{post.id}}_draft" 
+            <div class="form-field radio"> 
+                <input
+                    name="{{post.id}}_draft"
                     id="{{post.id}}_draft"
-                    value="draft" 
+                    value="draft"
                     type="radio"
                     ng-checked="post.status==='draft'"
                     ng-click="toggleRole('draft')">
@@ -14,11 +14,11 @@
             </div>
 
             <div class="form-field radio">
-                <input 
-                    name="{{post.id}}_everyone" 
-                    id="{{post.id}}_everyone" 
-                    value="" 
-                    type="radio" 
+                <input
+                    name="{{post.id}}_everyone"
+                    id="{{post.id}}_everyone"
+                    value=""
+                    type="radio"
                     ng-checked="post.status==='published' && post.published_to.length === 0"
                     ng-click="toggleRole('')">
                 <label for="{{post.id}}_everyone" translate>post.publish_for_everyone</label>
@@ -26,10 +26,10 @@
         </div>
         <div class="form-fieldgroup">
             <div class="form-field checkbox" ng-repeat="role in roles">
-                <input 
-                    id="{{role.id}}" 
-                    type="checkbox" 
-                    checklist-model="post.published_to" 
+                <input
+                    id="{{role.id}}"
+                    type="checkbox"
+                    checklist-model="post.published_to"
                     checklist-value="role.name"
                     ng-click="toggleRole(role.name)">
                 <label for="{{role.id}}">{{role.display_name}}</label>
