--- server/www/templates/posts/modify.html
@@ -1,4 +1,0 @@
-<main class="page-content layout-c" role="main">
-    <post-choose-form post="post" ng-if="! post.form.id" active-form="activeForm"></post-choose-form>
-    <post-editor post="post" ng-if="post.form.id" active-form="activeForm"></post-editor>
-</main>
