+++ server/www/templates/posts/modify/modify.html
@@ -1,0 +1,2 @@
+<post-choose-form post="post" ng-if="! post.form.id" active-form="activeForm"></post-choose-form>
+<post-editor post="post" ng-if="post.form.id" active-form="activeForm"></post-editor>
