+++ server/www/templates/posts/modify/post-status.html
@@ -1,0 +1,5 @@
+<div ng-show="allowedChangeStatus()" class="form-field select">
+  <publish-selector
+      toggle-role-func="publishPostTo(updatedPost)"
+      post="post" ></publish-selector>
+</div>
