--- server/www/templates/posts/post-editor.html
@@ -1,57 +1,0 @@
-´<form name="form" ng-submit="savePost()" novalidate>
-        <div class="page-header" ng-if="postMode !== 'bulk_data_import'">
-            <ul class="breadcrumb" ng-if="! isEdit">
-                <li>
-                    <a ng-click="goBack()" ng-href="/posts/create" translate>post.modify.add_a_post</a>
-                </li>
-            </ul>
-            <ul class="breadcrumb" ng-if="isEdit">
-                <li>
-                    <a ng-href="/posts/{{post.id}}" translate>{{post.title}}</a>
-                </li>
-            </ul>
-            <div ng-if="! isEdit" class="page-header-body">
-                <h1 class="page-title"><span class="icon-left fa-plus" translate translate-values="{ form_name: post.form.name}">post.add_post_type</span></h1>
-            </div>
-            <div ng-if="isEdit" class="page-header-body">
-                <h1 class="page-title"><span class="icon-left fa-pencil" translate translate-values="{form_name: post.form.name}">post.edit_post_type</span></h1>
-            </div>
-            <div class="page-header-actions">
-                <button type="submit" ng-disabled="! canSavePost() || saving_post" class="button-primary"><i class="fa fa-cog fa-spin" ng-show="saving_post"></i> <span translate>post.modify.save_post</span>
-                 </button>
-            </div>
-        </div>
-        <div class="page-header">
-            <div class="form-field title" ng-show="enableTitle">
-                <label class="input-label required"
-                    ng-class="{ 'error': form.title.$invalid && form.title.$dirty, 'success': !form.title.$invalid && form.title.$dirty }"
-                    for="title" translate>post.modify.form.title</label>
-                <div
-                    class="input-icon"
-                    ng-class="{ 'error': form.title.$invalid && form.title.$dirty }"
-                >
-                    <input id="title" ng-class="{ 'error': form.title.$invalid && form.title.$dirty }"
-                        name="title" type="text" placeholder="{{'post.modify.form.title' | translate }}" ng-model="post.title" required ng-minlength=2 ng-maxlength=150 class="form-control">
-                </div>
-                <p ng-repeat="(error, value) in form.title.$error"
-                    ng-show="form.title.$dirty"
-                    class="alert error"
-                    translate="{{'post.valid.title.' + error}}"></p>
-            </div>
-            <post-status
-                post="post"
-                stages="stages"
-            ></post-status>
-        </div>
-        <post-stages
-                form="form"
-                post="post"
-                stages="stages"
-                attributes="attributes"
-                visible-stage="visibleStage"
-        ></post-stages>
-        <div class="page-footer" ng-if="postMode !== 'bulk_data_import'">
-            <button type="submit" ng-disabled="! canSavePost() || saving_post" class="button-primary"><i class="fa fa-cog fa-spin" ng-show="saving_post"></i> <span translate>post.modify.save_post</span>
-            </button>
-        </div>
-</form>
