--- server/www/templates/posts/post-status.html
@@ -1,5 +1,0 @@
-<div ng-show="allowedChangeStatus()" class="form-field select">
-    <publish-selector
-        toggle-role-func="publishPostTo(updatedPost)"
-        post="post" ></publish-selector>
-</div>
